drivers: power: smb5-lib: add dcin aicl rerun
Rerun the dcin aicl after 30 seconds if we hit a uv
irq.  Also check that we are less than the min icl current
and not equal to it, since we will never hit it if our max
is not in 100 mA intervals.

Change-Id: Ieb890e110bee5ce809a2059a8906eedec6089661
Signed-off-by: Ryan Lattrel <ryanl@motorola.com>
Reviewed-on: https://gerrit.mot.com/1488904
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Ling Jin <lingjin@motorola.com>
Submit-Approved: Jira Key