kernel:drivers:usb:gadget f_fs.c add error checking
Some how we are calling ffs_ep_file_write_iter while disabling
the ffs driver. This causing it to block for a while. If you reconnect
in that time then it will force a disconnect after configuring.
Adding in this check means that if we have already disabled the function
we wont try to queue to a endpoint through read and write iter.

Change-Id: I00a477cd63e0b835e1b089b07c35a63577cbc305
Signed-off-by: Curtis Porter <porterc@motorola.com>
Reviewed-on: https://gerrit.mot.com/1489852
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Ling Jin <lingjin@motorola.com>
Submit-Approved: Jira Key